fix(lib/validate): move lib into server dir
We had a bug about the lib directory did not build with
the server side code. I think to move it into server dir
just a temporary solution as we probably would like to
use our libs at client side as well.

In the follow up patches we need to reorganise the
build process.